refactor: Move UiActionHandler and related imports to commonsKmm input package
